Add back button and abstract request tree construction
