+++ src/main/resources/assets/minecolonies/gui/windowpostbox.xml
@@ -24,12 +24,18 @@
             <!--Item icon -->
             <itemicon id="requestStack" size="16 16" pos="1 3"/>
             <image id="deliveryImage" source="minecraft:textures/misc/shadow.png" size="16 16" pos="1 3" visible="false"/>
+
             <!--Request name-->
             <buttonimage id="detail" size="16 16" pos="1 23" source="minecolonies:textures/gui/citizen/detail_button.png"/>
             <label id="shortDetail" size="50 12" pos="20 1" color="black"/>
             <label id="requester" size="50 12" pos="20 13" color="black"/>
+
+            <!-- Color green if possible? -->
+            <buttonimage id="fulfill" size="37 16" pos="20 23" textcolor="black"
+                         label="$(com.minecolonies.coremod.gui.citizen.requestsFulfill)" source="minecolonies:textures/gui/citizen/colonist_button_small.png"/>
+
             <!-- button to cancel the request-->
-            <buttonimage id="cancel" size="37 16" pos="50 23" textcolor="black"
+            <buttonimage id="cancel" size="37 16" pos="110 23" textcolor="black"
                          label="$(com.minecolonies.coremod.gui.citizen.requestsCancel)" source="minecolonies:textures/gui/citizen/colonist_button_small.png"/>
         </box>
     </list>
