+++ src/main/resources/assets/minecolonies/gui/windowrequestdetail.xml
@@ -17,6 +17,6 @@
     <label id="targetLocation" size="50 12" pos="42 44" color="black"/>
     <label id="resolver" size="50 12" pos="42 58" color="black"/>
 
-    <buttonimage id="back" size="37 16" pos="140 180" textcolor="black"
+    <buttonimage id="back" size="37 16" pos="80 100" textcolor="black"
                  label="Back" source="minecolonies:textures/gui/citizen/colonist_button_small.png"/>
 </window>
